createhdds: use user-mode networking for now
There's been an annoying bug on the production openQA server
for a while now: createhdds runs to refresh the upgrade base
images don't actually work most of the time, each attempt to
run virt-install fails. I finally got time to debug this a bit
today and it seems to be some kind of network issue: the VM has
no network access so the install doesn't work. I've no idea why
that's happening, but using user-mode networking seems to work
around it for now and shouldn't have any terrible side effects
even for openQA servers not affected by this problem. So let's
do this until we can pin down the real bug.