Allow 'stable' and 'current' as release for virt-install images
'stable' means 'all current stable releases', and is used for
desktop and server as we need to ensure we have those images
available for all stable releases (including the '-3' release
while it's not EOL) for update testing. (Currently, F24 update
tests are all failing as the images are missing).

'current' means 'the current stable release' - it's the same as
'-1', but just easier to understand. It's used for support.