Ensure domain is taken down when we exhaust retries
This was just an oversight - we don't destroy and undefine the
'createhdds' domain on the path where the final retry to create
an image times out, we just wipe the temp file and exit. So we
leave a qemu process sitting around until createhdds runs again
(and, on openQA production, exhausting swap space).

Signed-off-by: Adam Williamson <awilliam@redhat.com>