Explicitly enable firstboot in the desktop kickstarts
This ensures that gnome-initial-setup will run in user mode on
first login as 'test' user for F26+, as it did for <=F25 and as
the tests expected (we can now remove the workaround I added
because this stopped happening with F26). Bump the image version
so the images get rebuilt.