Don't produce ppc64 images for Fedora > 28
ppc64 (as opposed to ppc64le) was retired in Fedora 29. Tweaking
hdds.json to reflect this now is a bit awkward, so instead let's
leave it as-is, but add code to createhdds to not really include
a ppc64 image in the expected list if it's for F29 or later.
Once F28 goes EOL we can just drop all ppc64 from hdds.json and
lose this code.

Signed-off-by: Adam Williamson <awilliam@redhat.com>