Move i686 F31+ filter, drop ppc64 F29+ one
We previously did a filter like the one for 'i686 on F31+' for
'ppc64 on F29+', only in a different place which was actually a
better place. That filter is now unneeded as F28 went EOL, so
drop it...but move the i686 F31+ filter into the same place, as
it's better done there than in create().

Signed-off-by: Adam Williamson <awilliam@redhat.com>