Add more stuff to KDE image, build it for more releases
This is related to enabling update tests for KDE. While testing
that it transpired that the KDE disk image is missing various
stuff that is on the live image and systems installed from the
live image; this is because the kde-desktop-environment env
group doesn't by default include several groups and individual
packages that the live kickstart *does* include. So our current
disk image closely matches what you get if you do a network
install and pick the KDE env group and don't pick any option
groups, but that's not a very common thing to do, and it causes
problems for the tests (e.g. okular and firefox not being
there). I think it's reasonable to make the image resemble a
live install more closely; this is more convenient but I think
it's *also* more useful, as there are probably far more KDE
installs out there *with* these things than without them. More
KDE installs are probably done with the live than via network
install, and even KDE installs done via netinst may well have
these bits added subsequently.

Obviously, we bump the image version with this change, so we'll
need to update the scheduler and templates also.

Signed-off-by: Adam Williamson <awilliam@redhat.com>