Add plymouth-system-theme to server.ks (workaround #1933378)
The bug where the bootsplash doesn't always clear is really
playing havoc with F34 tests right now, and dealing with it by
workaround needles isn't going so well. I'm hoping installing a
real Plymouth theme might help, and in a quick trial run on lab
it seems to, so let's try this. The bug is still a bug, but it's
no good to us for update tests to randomly fail on it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>