Use '.qcow2' not '.img' as the extension for virt-install images
They're qcow2 images, so this is more correct, and will make
os-autoinst's new backing format autodetection work correctly.
We make the `check` function rename existing images, for
convenience.
NOTE: this requires a matching change to the templates in
os-autoinst, and you will want to run the check function to
rename all existing images to avoid wasting a ton of time
recreating them.

Signed-off-by: Adam Williamson <awilliam@redhat.com>